fatercal_apps/requirements.txt
Django==2.2.24
django-jet==1.0.8
psycopg2==2.8.6
django-ajax-selects==1.7.1
numpy==1.17.4
djangorestframework==3.11.2
Markdown==3.2.1